% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Compute GCD using the Euclidean algorithm}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{First input scalar numeric value.}

\item{y}{Second input scalar numeric value.}
}
\value{
The greatest common divisor between x and y, a numeric value.
}
\description{
´euclidean()´ calculates the greatest common devisor (GCD),
of two scalar values.
}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
\examples{
euclidean(252,105)
}
