% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Example graph from Dijkstra Wiki}
\format{
An object of class \code{data.frame} with 18 rows and 3 columns.
}
\source{
<https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm>
}
\usage{
wiki_graph
}
\description{
\describe{
\item{v1}{Starting vertex of edge}
\item{v2}{Ending vertex of edge}
\item{w}{Weight of the edge}
}
}
\keyword{datasets}
